Stabilize grid-match jumpiness
The "revert" method caused grid columns to jump around during window resize. The solution is to get the height of each column by contents, not by the column itself, so that maximum column height can be determined without having to revert the height of each column, thus stabilizing the jumpy behavior.