filter custom directory for IDE-files
While indexing the theme a theme `.idea` is indexed because of the hidden folder PHPstorm creates there. I expanded the existing filter for `blank` and `.git` with `.idea` and refactored the code so more filters can be added easily later.
Would you consider merging this, so I don;t have to edit the Grunt in a local branch? Thanks